Move the components folder outside of the app folder to keep the app folder only for the routes.
