--- app/components/styles/menu.css
@@ -1,28 +1,0 @@
-/* Mobile menu */
-
-.hamburger div {
-  transition: 0.5s;
-}
-
-.hamburger:hover div {
-  width: 100%;
-  transition: all 0.3s;
-}
-
-.line-1 {
-  width: 40%;
-}
-
-.line-2 {
-  width: 80%;
-}
-
-.line-3 {
-  width: 50%;
-}
-
-/* Desktop menu */
-
-.desktopMenu {
-  background-color: rgba(255, 255, 255, 0.1);
-}
