--- app/js/data/home.js
@@ -1,38 +1,0 @@
-export const hero = [
-  {
-    subTitle: "Welcome to my digital space where innovation meets expertise.",
-    title: "Experienced Full-stack",
-    title2: "Software Engineer",
-    desc1:
-      "From frontend to backend mastery, I thrive in turning complex ideas into elegant, user-centric applications.",
-    desc2:
-      "Let's embark on a journey to explore the endless possibilities of code...",
-  },
-];
-
-export const infosHero = [
-  {
-    text: "Email",
-    content: "jeanemmanuelcadet@gmail.com",
-  },
-  {
-    text: "Phone",
-    content: "+1 (267) 529 6040",
-  },
-  {
-    text: "Work location",
-    content: "Remotely",
-  },
-];
-
-export const aboutme = [
-  {
-    title: "I'm Jean Emmanuel Cadet",
-    paragraph:
-      "I'm a Full-stack software engineer who specializes in using Javascript, React/Redux, Next.js, Ruby and Ruby on Rails to build software solutions for small and larger businesses.",
-    paragraph2:
-      "I'm currently completing coursework at Microverse, a full-time remote software development program that allows devs from all over the world to use pair programming and project building to teach development.",
-    paragraph3:
-      "I've spent the last several months working remotely and asynchronously with people across different time zones.",
-  },
-];
