--- app/js/fonts.js
@@ -1,12 +1,0 @@
-import { Nunito, Rubik } from "next/font/google";
-
-export const nunito = Nunito({
-  weight: ['400', '700', '900'],
-  subsets: ['latin'],
-  display: 'swap',
-});
-
-export const rubik = Rubik({
-  subsets: ['latin'],
-  display: 'swap',
-});
